refactor: 포인트 검증 서비스(역할) 분리
