+++ src/main/java/io/hhplus/tdd/point/DefaultPointService.java
@@ -9,18 +9,18 @@ import java.util.List;
 @Service
 public class DefaultPointService implements PointService {
 
-    private static final long USER_MAX_POINT = 100_000_000L;
-    private static final int ONCE_CHARGE_MAX_POINT = 10_000;
-
     private final UserPointTable userPointTable;
     private final PointHistoryTable pointHistoryTable;
+    private final PointValidationService pointValidationService;
 
     public DefaultPointService(
             UserPointTable userPointTable,
-            PointHistoryTable pointHistoryTable
+            PointHistoryTable pointHistoryTable,
+            PointValidationService pointValidationService
     ) {
         this.userPointTable = userPointTable;
         this.pointHistoryTable = pointHistoryTable;
+        this.pointValidationService = pointValidationService;
     }
 
     @Override
@@ -40,17 +40,10 @@ public class DefaultPointService implements PointService {
      * 3. 0 이하의 금액은 충전할 수 없다.
      */
     public UserPoint charge(long userId, long amount) {
-        if (amount <= 0) {
-            throw new IllegalArgumentException("0 이하의 금액을 충전할 수 없습니다.");
-        }
-        if (amount > ONCE_CHARGE_MAX_POINT) {
-            throw new IllegalArgumentException("한번에 충전 가능한 금액을 초과했습니다.");
-        }
+        pointValidationService.validateCharge(userId, amount);
+
         UserPoint userPoint = userPointTable.selectById(userId);
         long amountToSave = userPoint.point() + amount;
-        if (amountToSave > USER_MAX_POINT) {
-            throw new IllegalArgumentException("최대 잔고를 초과하여 충전할 수 없습니다.");
-        }
 
         UserPoint savedUserPoint = userPointTable.insertOrUpdate(userId, amountToSave);
         pointHistoryTable.insert(userId, amount, TransactionType.CHARGE, System.currentTimeMillis());
@@ -65,14 +58,10 @@ public class DefaultPointService implements PointService {
      */
     @Override
     public UserPoint use(long userId, long amount) {
-        if (amount <= 0) {
-            throw new IllegalArgumentException("0 이하의 금액을 사용할 수 없습니다.");
-        }
+        pointValidationService.validateUse(userId, amount);
+
         UserPoint userPoint = userPointTable.selectById(userId);
         long amountToSave = userPoint.point() - amount;
-        if (amountToSave < 0) {
-            throw new IllegalArgumentException("잔액이 부족합니다.");
-        }
 
         UserPoint savedUserPoint = userPointTable.insertOrUpdate(userId, amountToSave);
         pointHistoryTable.insert(userId, amount, TransactionType.USE, System.currentTimeMillis());
